Move plugin and split sync helpers
